feat: Provide `replaceNonTextChars`
Some xml documents (purposely) contain characters that are not visible and make it hard to reason about a test result.
Those characters also cause git to think a file is binary,
when the test output is being committed (e.g. in a jest snapshot).
By replacing them with a visual replacement using `codePointAt` (e.g. replacing NUL/`&#0;` with `{!0!}`,
it's makes those more obvious.